Use unique scope Id and reorganize code
